requirements.txt
bcrypt==3.1.7
cached-property==1.5.1
certifi==2019.9.11
cffi==1.13.0
chardet==3.0.4
coverage==4.5.4
cryptography==2.8
Django==2.2.5
django-crontab==0.7.1
django-ipware==2.1.0
django-material-admin==1.6.7
django-python3-ldap==0.11.2
djangorestframework==3.10.3
docker==3.7.3
docker-compose==1.24.1
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
gunicorn==19.9.0
idna==2.7
jsonschema==2.6.0
ldap3==2.5
paramiko==2.6.0
psycopg2-binary==2.8.3
pyasn1==0.4.4
pycparser==2.19
PyNaCl==1.3.0
python-decouple==3.1
pytz==2019.2
PyYAML==3.13
requests==2.20.1
six==1.12.0
sqlparse==0.3.0
texttable==0.9.1
urllib3==1.24.3
websocket-client==0.56.0
